Streamlit/requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-together
langchain-google-genai
pydantic
pillow
tiktoken

python-multipart
pandas
numpy
uvicorn
starlette
python-dotenv
pydantic-settings
sentence-transformers

